requirements.txt
Flask
pandas
plotly
mysql-connector-python
scikit-learn
